+++ moPepGen/cli/summarize_fasta.py
@@ -9,6 +9,7 @@ from pathlib import Path
 import sys
 from typing import IO
 import matplotlib.pyplot as plt
+from moPepGen import logger
 from moPepGen.cli import common
 from moPepGen.aa.PeptidePoolSummarizer import PeptidePoolSummarizer
 
@@ -150,19 +151,26 @@ def summarize_fasta(args:argparse.Namespace) -> None:
     for gvf in args.gvf:
         with open(gvf, 'rt') as handle:
             summarizer.update_label_map(handle)
+        logger(f"GVF file used: {gvf}")
 
     summarizer.append_order_internal_sources()
 
     with open(args.variant_peptides, 'rt') as handle:
         summarizer.load_database(handle)
 
+    logger(f"Variant FASTA loaded: {args.variant_peptides}")
+
     if args.noncoding_peptides:
         with open(args.noncoding_peptides, 'rt') as handle:
             summarizer.load_database(handle)
+        logger(f"Noncoding FASTA loaded: {args.noncoding_peptides}")
 
     if args.alt_translation_peptides:
         with open(args.alt_translation_peptides, 'rt') as handle:
             summarizer.load_database(handle)
+        logger(f"Alternative Translation FASTA loaded: {args.alt_translation_peptides}")
+
+    logger("Start summarizing..")
 
     summarizer.count_peptide_source(anno, args.cleavage_rule)